Split up service
